Replace pinball approach with crokinole-inspired approach
Using a cylinder shape was necessary to get the physics to be predictable.
A generated collision mesh did not work, and so putting a hole in the middle
was untenable. It might be possible if the mesh were imported into blender
and a collision shape hand-crafted, but even CSG collisions were inadequate
to get the smooth sailing we see in this version.